Add MANIFEST.in to include all data files
This is needed because when running with --debug it tried to find the
templates in the package data directory instead of the source directory.

Setuptools claims that "include_package_data" will include everything
that is under version control, but the only supported version control
systems are CVS and SVN.

I thought setuptools was created after those stopped being popular.